Fix logout packets being incorrect for wrath, add TBC
