+++ wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm
@@ -1,3 +1,5 @@
+#tag_all versions "1.12 2 3";
+
 enum LogoutResult : u32 {
     SUCCESS = 0;
     FAILURE_IN_COMBAT = 1;
@@ -5,31 +7,11 @@ enum LogoutResult : u32 {
         comment = "vmangos checks for aura 9454. Has FIXME - Need the correct value.";
     }
     FAILURE_JUMPING_OR_FALLING = 3;
-} {
-    versions = "1.12";
 }
 
 enum LogoutSpeed : u8 {
     DELAYED = 0;
     INSTANT = 1;
-} {
-    versions = "1.12";
-}
-
-enum LogoutResult : u16 {
-    SUCCESS = 0;
-    FAILURE_IN_COMBAT = 1;
-    FAILURE_FROZEN_BY_GM = 2;
-    FAILURE_JUMPING_OR_FALLING = 3;
-} {
-    versions = "3";
-}
-
-enum LogoutSpeed : u16 {
-    DELAYED = 0;
-    INSTANT = 1;
-} {
-    versions = "3";
 }
 
 smsg SMSG_LOGOUT_RESPONSE = 0x4C {
@@ -38,7 +20,6 @@ smsg SMSG_LOGOUT_RESPONSE = 0x4C {
 } {
     description = "Reply to [CMSG_LOGOUT_REQUEST].";
     comment = "The client expects to get an [SMSG_LOGOUT_COMPLETE] when logout is complete.";
-    paste_versions = "1.12 3";
 }
 
 test SMSG_LOGOUT_RESPONSE {
@@ -49,7 +30,4 @@ test SMSG_LOGOUT_RESPONSE {
      0x4C, 0x00, /* opcode */
      0x00, 0x00, 0x00, 0x00, /* reason: SUCCESS */
      0x01, /* speed = INSTANT */
-] {
-    versions = "1.12";
-}
-
+]
