+++ wow_world_messages/src/world/shared/mod.rs
@@ -166,6 +166,8 @@ pub mod language_tbc_wrath;
 pub mod language_vanilla_vanilla;
 pub mod list_inventory_item_tbc_wrath;
 pub mod log_format_vanilla_tbc_wrath;
+pub mod logout_result_vanilla_tbc_wrath;
+pub mod logout_speed_vanilla_tbc_wrath;
 pub mod msg_auction_hello_client_vanilla_tbc_wrath;
 pub mod msg_auction_hello_server_vanilla_tbc;
 pub mod msg_battleground_player_positions_client_vanilla_tbc_wrath;
@@ -251,6 +253,7 @@ pub mod smsg_log_xpgain_vanilla_tbc;
 pub mod smsg_login_settimespeed_vanilla_tbc;
 pub mod smsg_logout_cancel_ack_vanilla_tbc_wrath;
 pub mod smsg_logout_complete_vanilla_tbc_wrath;
+pub mod smsg_logout_response_vanilla_tbc_wrath;
 pub mod smsg_notification_vanilla_tbc_wrath;
 pub mod smsg_petition_show_signatures_vanilla_tbc_wrath;
 pub mod smsg_petition_sign_results_tbc_wrath;
