+++ wow_world_messages/src/world/tbc/logout_result.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::logout_result_vanilla_tbc_wrath::LogoutResult;
+
