+++ wow_world_messages/src/world/tbc/logout_speed.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::logout_speed_vanilla_tbc_wrath::LogoutSpeed;
+
