+++ wow_world_messages/src/world/tbc/mod.rs
@@ -409,6 +409,10 @@ pub(crate) mod list_inventory_item;
 pub use list_inventory_item::*;
 pub(crate) mod log_format;
 pub use log_format::*;
+pub(crate) mod logout_result;
+pub use logout_result::*;
+pub(crate) mod logout_speed;
+pub use logout_speed::*;
 pub(crate) mod map;
 pub use map::*;
 pub(crate) mod movement_block;
@@ -740,6 +744,8 @@ pub(crate) mod smsg_logout_cancel_ack;
 pub use smsg_logout_cancel_ack::*;
 pub(crate) mod smsg_logout_complete;
 pub use smsg_logout_complete::*;
+pub(crate) mod smsg_logout_response;
+pub use smsg_logout_response::*;
 pub(crate) mod smsg_new_world;
 pub use smsg_new_world::*;
 pub(crate) mod smsg_notification;
