+++ wow_world_messages/src/world/tbc/smsg_logout_response.rs
@@ -1,0 +1,2 @@
+pub use crate::shared::smsg_logout_response_vanilla_tbc_wrath::SMSG_LOGOUT_RESPONSE;
+
