+++ wowm_language/src/docs/logoutresult.md
@@ -1,10 +1,10 @@
 # LogoutResult
 
-## Client Version 1.12
+## Client Version 1.12, Client Version 2, Client Version 3
 
 ### Wowm Representation
 
-Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm:1`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm#L1).
+Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm:3`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm#L3).
 
 ```rust,ignore
 enum LogoutResult : u32 {
@@ -22,35 +22,6 @@ The basic type is `u32`, a 4 byte (32 bit) little endian integer.
 | `SUCCESS` | 0 (0x00) |  |  |
 | `FAILURE_IN_COMBAT` | 1 (0x01) |  |  |
 | `FAILURE_FROZEN_BY_GM` | 2 (0x02) |  | vmangos checks for aura 9454. Has FIXME - Need the correct value. |
-| `FAILURE_JUMPING_OR_FALLING` | 3 (0x03) |  |  |
-
-Used in:
-* [SMSG_LOGOUT_RESPONSE](smsg_logout_response.md)
-
-# LogoutResult
-
-## Client Version 3
-
-### Wowm Representation
-
-Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm:19`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm#L19).
-
-```rust,ignore
-enum LogoutResult : u16 {
-    SUCCESS = 0;
-    FAILURE_IN_COMBAT = 1;
-    FAILURE_FROZEN_BY_GM = 2;
-    FAILURE_JUMPING_OR_FALLING = 3;
-}
-```
-### Type
-The basic type is `u16`, a 2 byte (16 bit) little endian integer.
-### Enumerators
-| Enumerator | Value  | Description | Comment |
-| --------- | -------- | ----------- | ------- |
-| `SUCCESS` | 0 (0x00) |  |  |
-| `FAILURE_IN_COMBAT` | 1 (0x01) |  |  |
-| `FAILURE_FROZEN_BY_GM` | 2 (0x02) |  |  |
 | `FAILURE_JUMPING_OR_FALLING` | 3 (0x03) |  |  |
 
 Used in:
