+++ wowm_language/src/docs/logoutspeed.md
@@ -1,6 +1,6 @@
 # LogoutSpeed
 
-## Client Version 1.12
+## Client Version 1.12, Client Version 2, Client Version 3
 
 ### Wowm Representation
 
@@ -14,31 +14,6 @@ enum LogoutSpeed : u8 {
 ```
 ### Type
 The basic type is `u8`, a 1 byte (8 bit) integer.
-### Enumerators
-| Enumerator | Value  | Description | Comment |
-| --------- | -------- | ----------- | ------- |
-| `DELAYED` | 0 (0x00) |  |  |
-| `INSTANT` | 1 (0x01) |  |  |
-
-Used in:
-* [SMSG_LOGOUT_RESPONSE](smsg_logout_response.md)
-
-# LogoutSpeed
-
-## Client Version 3
-
-### Wowm Representation
-
-Autogenerated from `wowm` file at [`wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm:28`](https://github.com/gtker/wow_messages/tree/main/wow_message_parser/wowm/world/login_logout/smsg_logout_response.wowm#L28).
-
-```rust,ignore
-enum LogoutSpeed : u16 {
-    DELAYED = 0;
-    INSTANT = 1;
-}
-```
-### Type
-The basic type is `u16`, a 2 byte (16 bit) little endian integer.
 ### Enumerators
 | Enumerator | Value  | Description | Comment |
 | --------- | -------- | ----------- | ------- |
